Refactor API tests into package
