--- backend/apps/api/tests.py
@@ -1,429 +1,0 @@
-from unittest.mock import patch, call
-from django.test import TestCase, Client
-
-from apps.api.models import PlayerIdInfo, TeamIdInfo, Venue
-
-
-class ScheduleApiTests(TestCase):
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_schedule_endpoint(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-        mock_client.get_schedule_for_date_range.return_value = [
-            {
-                'date': '2025-08-18',
-                'games': [
-                    {
-                        'gameDate': '2025-08-18T18:20:00Z',
-                        'teams': {
-                            'away': {'team': {'name': 'Away Team', 'id': 1}},
-                            'home': {'team': {'name': 'Home Team', 'id': 2}},
-                        },
-                    }
-                ]
-            }
-        ]
-        mock_client.get_team_spot_url.return_value = 'logo-url'
-        client = Client()
-        response = client.get('/api/schedule/', {'date': '2025-08-18'})
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertEqual(len(data), 1)
-        game = data[0]['games'][0]
-        self.assertEqual(game['teams']['home']['team']['logo_url'], 'logo-url')
-        self.assertEqual(game['teams']['away']['team']['logo_url'], 'logo-url')
-
-
-class GameDataApiTests(TestCase):
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_game_data_endpoint(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-        mock_client.get_game_live_feed.return_value = {
-            'teams': {
-                'home': {'team': {'name': 'Home'}, 'score': 5},
-                'away': {'team': {'name': 'Away'}, 'score': 3},
-            },
-            'home_team_data': {'id': 1},
-            'away_team_data': {'id': 2},
-            'liveData': {
-                'boxscore': {'info': [{'label': 'Att', 'value': '10,000'}]}
-            },
-        }
-        mock_client.get_team_spot_url.return_value = 'logo-url'
-        client = Client()
-        response = client.get('/api/games/123/')
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertEqual(data['teams']['home']['team']['name'], 'Home')
-        self.assertEqual(data['teams']['away']['score'], 3)
-        # boxscore data should be present in liveData
-        self.assertEqual(
-            data['liveData']['boxscore']['info'][0]['label'],
-            'Att',
-        )
-
-
-class StandingsApiTests(TestCase):
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_standings_endpoint(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-        mock_client.get_standings_data.return_value = {
-            'records': [
-                {
-                    'league': {'name': 'American League'},
-                    'division': {'name': 'East'},
-                    'teamRecords': [
-                        {'team': {'id': 1, 'name': 'Team A'}, 'wins': 10, 'losses': 5},
-                        {'team': {'id': 2, 'name': 'Team B'}, 'wins': 8, 'losses': 7},
-                    ],
-                }
-            ]
-        }
-        client = Client()
-        response = client.get('/api/standings/')
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertIn('records', data)
-        self.assertEqual(len(data['records'][0]['teamRecords']), 2)
-
-
-class NewsApiTests(TestCase):
-    def test_news_endpoint(self):
-        client = Client()
-        response = client.get('/api/news/')
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertTrue(len(data) > 0)
-        self.assertIn('title', data[0])
-
-
-class PlayerHeadshotApiTests(TestCase):
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_player_headshot_endpoint(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-        mock_client.fetch_player_headshot.return_value = b'image-bytes'
-
-        PlayerIdInfo.objects.create(
-            id=1,
-            key_mlbam='123',
-            name_first='Test',
-            name_last='Player',
-        )
-
-        client = Client()
-        response = client.get('/api/players/1/headshot/')
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.content, b'image-bytes')
-        self.assertEqual(response['Content-Type'], 'image/png')
-        mock_client.fetch_player_headshot.assert_called_once_with(123)
-
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_player_headshot_endpoint_accepts_mlbam_id(self, mock_client_cls):
-        """The endpoint should work when passed a raw MLBAM id."""
-        mock_client = mock_client_cls.return_value
-        mock_client.fetch_player_headshot.return_value = b'image-bytes'
-
-        client = Client()
-        response = client.get('/api/players/456/headshot/')
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.content, b'image-bytes')
-        self.assertEqual(response['Content-Type'], 'image/png')
-        mock_client.fetch_player_headshot.assert_called_once_with(456)
-
-
-class PlayerInfoApiTests(TestCase):
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_player_info_endpoint(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-        mock_client.fetch_player_info.return_value = {
-            'currentTeam': {'id': 111, 'name': 'Team A'},
-            'primaryPosition': {'name': 'Pitcher'},
-            'birthDate': '1990-01-01',
-            'birthCity': 'Some City',
-            'birthStateProvince': 'CA',
-            'birthCountry': 'USA',
-            'height': "6'2\"",
-            'weight': 200,
-            'batSide': {'description': 'Right'},
-            'pitchHand': {'description': 'Left'},
-        }
-
-        PlayerIdInfo.objects.create(
-            id=1,
-            key_mlbam='123',
-            name_first='Test',
-            name_last='Player',
-        )
-
-        client = Client()
-        response = client.get('/api/players/1/')
-
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertEqual(data['team_id'], 111)
-        self.assertEqual(data['team_name'], 'Team A')
-        self.assertEqual(data['position'], 'Pitcher')
-        self.assertEqual(data['birth_date'], '1990-01-01')
-        self.assertEqual(data['birth_place'], 'Some City, CA, USA')
-        self.assertEqual(data['height'], "6'2\"")
-        self.assertEqual(data['weight'], 200)
-        self.assertEqual(data['bat_side'], 'Right')
-        self.assertEqual(data['throw_side'], 'Left')
-        mock_client.fetch_player_info.assert_called_once_with(123)
-
-
-class PlayerStatsApiTests(TestCase):
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_player_stats_endpoint(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-        mock_client.fetch_player_stats_career.side_effect = [
-            {'stats': ['bat']},
-            {'stats': ['pitch']},
-        ]
-
-        PlayerIdInfo.objects.create(
-            id=1,
-            key_mlbam='123',
-            name_first='Test',
-            name_last='Player',
-        )
-
-        client = Client()
-        response = client.get('/api/players/1/stats/')
-
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertEqual(data['batting']['stats'], ['bat'])
-        self.assertEqual(data['pitching']['stats'], ['pitch'])
-        mock_client.fetch_player_stats_career.assert_has_calls([
-            call(123, group='batting'),
-            call(123, group='pitching'),
-        ])
-
-
-class PlayerSearchApiTests(TestCase):
-    def setUp(self):
-        # Create more than 10 players to ensure the limit works
-        for i in range(11):
-            PlayerIdInfo.objects.create(
-                name_first='Test',
-                name_last=f'Player{i}',
-                key_mlbam=str(100 + i),
-            )
-        PlayerIdInfo.objects.create(
-            name_first='Foo',
-            name_last='Bar',
-            key_mlbam='123.0',
-        )
-
-    def test_player_search_case_insensitive_and_limited(self):
-        client = Client()
-        # Case-insensitive search and limit
-        response = client.get('/api/players/', {'q': 'test'})
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertEqual(len(data), 10)
-        self.assertTrue(all('Test Player' in p['name_full'] for p in data))
-
-        # Ensure key_mlbam is sanitized and search is case-insensitive
-        response = client.get('/api/players/', {'q': 'foo'})
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertEqual(len(data), 1)
-        self.assertEqual(data[0]['key_mlbam'], '123')
-
-
-class TeamSearchApiTests(TestCase):
-    def setUp(self):
-        TeamIdInfo.objects.create(id=1, full_name='Red Sox', mlbam_team_id=111)
-        TeamIdInfo.objects.create(id=2, full_name='Blue Jays', mlbam_team_id=222)
-
-    def test_team_search_returns_results(self):
-        client = Client()
-        response = client.get('/api/teams/', {'q': 'red'})
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertEqual(len(data), 1)
-        self.assertEqual(data[0]['full_name'], 'Red Sox')
-        self.assertEqual(data[0]['mlbam_team_id'], '111')
-
-
-class TeamInfoApiTests(TestCase):
-    def setUp(self):
-        TeamIdInfo.objects.create(
-            id=1,
-            full_name='Red Sox',
-            mlbam_team_id=111,
-            location_name='Boston',
-            abbrev='BOS'
-        )
-        Venue.objects.create(mlbam_id=10, name='Fenway Park', link='link', active=True, season=2024)
-
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_team_info_returns_team(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-        mock_client.fetch_team.return_value = {'venue': {'id': 10}}
-
-        client = Client()
-        response = client.get('/api/teams/111/')
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertEqual(data['full_name'], 'Red Sox')
-        self.assertEqual(data['mlbam_team_id'], '111')
-        self.assertEqual(data['location_name'], 'Boston')
-        self.assertEqual(data['abbrev'], 'BOS')
-        self.assertEqual(data['venue_id'], '10')
-        self.assertEqual(data['venue']['name'], 'Fenway Park')
-
-    def test_team_info_not_found(self):
-        client = Client()
-        response = client.get('/api/teams/999/')
-        self.assertEqual(response.status_code, 404)
-
-
-class TeamLogoApiTests(TestCase):
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_team_logo_endpoint(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-        mock_client.get_team_logo_url.return_value = 'logo-url'
-
-        TeamIdInfo.objects.create(id=1, mlbam_team_id=555, full_name='Team A')
-
-        client = Client()
-        response = client.get('/api/teams/1/logo/')
-
-        self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.content, b'logo-url')
-        self.assertEqual(response['Content-Type'], 'text/plain')
-        mock_client.get_team_logo_url.assert_called_once_with(555)
-
-
-class TeamRecordApiTests(TestCase):
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_team_record_endpoint(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-        mock_client.get_team_record_for_season.return_value = {
-            'wins': 10,
-            'losses': 5,
-            'divisionRank': '1',
-            'streak': {
-                'streakType': 'wins',
-                'streakNumber': 3,
-                'streakCode': 'W3'
-            }
-        }
-
-        TeamIdInfo.objects.create(id=1, mlbam_team_id=555, full_name='Team A')
-
-        client = Client()
-        response = client.get('/api/teams/1/record/', {'season': '2025'})
-
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertEqual(data['wins'], 10)
-        mock_client.get_team_record_for_season.assert_called_once_with(555, 2025)
-
-
-class TeamRecentScheduleApiTests(TestCase):
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_team_recent_schedule_endpoint(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-        mock_client.get_recent_schedule_for_team.return_value = {
-            'id': 555,
-            'previousGameSchedule': {'dates': []},
-            'nextGameSchedule': {'dates': []}
-        }
-
-        TeamIdInfo.objects.create(id=1, mlbam_team_id=555, full_name='Team A')
-
-        client = Client()
-        response = client.get('/api/teams/1/recent_schedule/')
-
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertEqual(data['id'], 555)
-        mock_client.get_recent_schedule_for_team.assert_called_once_with(555)
-
-    def test_team_recent_schedule_team_not_found(self):
-        client = Client()
-        response = client.get('/api/teams/1/recent_schedule/')
-        self.assertEqual(response.status_code, 404)
-
-
-class TeamRosterApiTests(TestCase):
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_team_roster_endpoint(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-        mock_client.get_team_roster.return_value = {
-            'roster': [
-                {
-                    'id': 1,
-                    'fullName': 'Test Player',
-                    'primaryPosition': {'abbreviation': 'P'},
-                    'stats': {'gamesPlayed': 10},
-                }
-            ]
-        }
-
-        TeamIdInfo.objects.create(id=1, mlbam_team_id=555, full_name='Team A')
-
-        client = Client()
-        response = client.get('/api/teams/1/roster/')
-
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertEqual(len(data['roster']), 1)
-        mock_client.get_team_roster.assert_called_once_with(555)
-
-    def test_team_roster_team_not_found(self):
-        client = Client()
-        response = client.get('/api/teams/1/roster/')
-        self.assertEqual(response.status_code, 404)
-
-
-class LeagueLeadersApiTests(TestCase):
-    @patch('apps.api.views.UnifiedDataClient')
-    def test_league_leaders_endpoint(self, mock_client_cls):
-        mock_client = mock_client_cls.return_value
-
-        import pandas as pd
-
-        bat_df = pd.DataFrame(
-            {
-                'xMLBAMID': [1, 2],
-                'Name': ['<b>Player A</b>', 'Player B'],
-                'PA': [100, 120],
-                'HR': [10, 20],
-                'AVG': [0.300, 0.350],
-                'RBI': [30, 40],
-                'SB': [5, 10],
-                'SLG': [0.500, 0.600],
-                'OPS': [0.800, 0.900],
-            }
-        )
-        pit_df = pd.DataFrame(
-            {
-                'xMLBAMID': [3, 4],
-                'Name': ['Pitcher C', 'Pitcher D'],
-                'Pos': ['P', 'P'],
-                'IP': [50, 60],
-                'ERA': [2.5, 3.5],
-                'SO': [100, 80],
-                'W': [8, 10],
-                'SV': [1, 5],
-            }
-        )
-
-        mock_client.fetch_batting_leaderboards.return_value = bat_df
-        mock_client.fetch_pitching_leaderboards.return_value = pit_df
-
-        client = Client()
-        response = client.get('/api/leaders/')
-
-        self.assertEqual(response.status_code, 200)
-        data = response.json()
-        self.assertIn('HR', data['batting'])
-        self.assertEqual(data['batting']['HR'][0]['id'], '2')
-        self.assertIn('ERA', data['pitching'])
-        self.assertEqual(data['pitching']['ERA'][0]['id'], '3')
